requirements.txt
# Core dependencies
edge-tts>=6.1.0
pydub>=0.25.1

# For audio processing
aiohttp>=3.8.0
numpy>=1.21.0

# For Google TTS
google-cloud-texttospeech>=2.14.0
google-auth>=2.15.0
google-api-core>=2.11.0

# For script parsing
markdown>=3.4.1
beautifulsoup4>=4.11.1
lxml>=4.9.1

# For CLI and rich output
typer>=0.9.0
rich>=13.0.0
PyYAML>=6.0

# For development and testing
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-mock>=3.10.0
